requirements.txt
strands-agents
strands-agents-tools
litellm
python-dotenv
